pyproject.toml
[tool.poetry]
name = "flex-datafortelling"
version = "0.1.0"
description = ""
authors = ["NAV-IT"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
nbformat = "^5.9.2"
nbclient = "^0.9.0"
ipython = "^8.21.0"
ipykernel = "^6.29.2"
python-dotenv = "^1.0.1"
matplotlib = "^3.8.3"
google-cloud-bigquery = "^3.17.2"
pandas = "^2.2.0"
db-dtypes = "^1.2.0"
pyyaml = "^6.0.1"
seaborn = "^0.13.2"
rich = "^13.7.1"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
